Review my changes and let us go over the creation of webpages in django.
I attempted to create another registration page just for the pharmacist but I am still confused about how Django is operating as some of the code in the HTML is missing compared to what is on the site. I worked on making it look cleaner with buttons, I just need to understand Django better so I can make other web pages.